create home screen and ui - separate home and market page

+++ app/src/main/java/com/finance/trade_learn/models/create_new_model_for_tem_history/newModelForItemMyCoins.kt
@@ -4,7 +4,7 @@ import java.math.BigDecimal
 
 data class NewModelForItemHistory(
     var CoinName: String,
-    var CoinAmount: BigDecimal,
+    var CoinAmount: Double,
     var Total: BigDecimal,
     val Image: String
